proyecto-clasificacion-basura/requirements.txt
tensorflow>=2.8.0
opencv-python
matplotlib
seaborn
scikit-learn
pandas
numpy
split-folders
pyyaml
jupyter
albumentations
scikit-image
tqdm
Pillow
flask
gunicorn
tabulate